requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
fasteners==0.14.1
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.2
httplib2==0.14.0
idna==2.8
itsdangerous==2.0.1
Jinja2==3.0.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.4.25
ssh-import-id==5.10
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==2.0.2
WTForms==2.3.3
xkit==0.0.0
